requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
pymongo[srv]==3.6
pymongo
certifi
dotenv
Pyaml
-e .

# -e . is trigger setup.py file